+++ lib/docusign_rest/client.rb
@@ -669,7 +669,9 @@ module DocusignRest
     # customFields  - (Optional) A hash of listCustomFields and textCustomFields.
     #                 Each contains an array of corresponding customField hashes.
     #                 For details, please see: http://bit.ly/1FnmRJx
-    # headers       - Allows a client to pass in some
+    # headers       - Allows a client to pass in some headers
+    # web_sign      - (Optional) If true, the signer is allowed to print the
+    #                 document and sign it on paper. False if not defined.
     #
     # Returns a JSON parsed response object containing:
     #   envelopeId     - The envelope's ID
@@ -680,7 +682,7 @@ module DocusignRest
       ios = create_file_ios(options[:files])
       file_params = create_file_params(ios)
 
-      post_body = {
+      post_hash = {
         emailBlurb:   "#{options[:email][:body] if options[:email]}",
         emailSubject: "#{options[:email][:subject] if options[:email]}",
         documents: get_documents(ios),
@@ -691,7 +693,9 @@ module DocusignRest
         eventNotification: get_event_notification(options[:event_notification]),
         status: "#{options[:status]}",
         customFields: options[:custom_fields]
-      }.to_json
+      }
+      post_hash[:enableWetSign] = options[:wet_sign] if options.has_key? :web_sign
+      post_body = post_hash.to_json
 
       uri = build_uri("/accounts/#{acct_id}/envelopes")