+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Support routingOrder option when generating signers (Guillaume Dott)
 * Support arbitrary parameters to DocusignRest::Client#get_combined_document_from_envelope (Coley Brown)
 * Support event notifications in DocusignRest::Client#get_combined_document_from_envelope (Maxime Orefice)
+* Support wet_sign option on DocusignRest::Client#create_envelope_from_document (Sergio Cambra)
 
 ### Misc:
 * Replace monkeypatch with argument usage (Jean-Philippe Moal)